renamed to commit phase.
